SPEC,FIX: Restored find & replace error in normal spec

+++ spec/rand/f32/rnorm_spec.cr
@@ -11,7 +11,7 @@ describe Alea do
           return_type: Float32,
           types: [Int8, Int16, Int32, Int64, Int128,
                   UInt8, UInt16, UInt32, UInt64, UInt128,
-                  Float32, Float32,
+                  Float32, Float64,
           ]
         )
